@@ -16,27 +16,26 @@
   "support": {
     "source": "https://github.com/imbo/behat-api-extension",
     "docs": "http://behat-api-extension.readthedocs.io/",
-    "issues": "https://github.com/imbo/behat-api-extension/issues",
-    "irc": "irc://irc.freenode.net/imbo"
+    "issues": "https://github.com/imbo/behat-api-extension/issues"
   },
   "require": {
-    "php": "^7.4 || ^8.0",
+    "php": ">=8.1",
     "ext-json": "*",
     "beberlei/assert": "^3.3",
     "behat/behat": "^3.8",
-    "firebase/php-jwt": "^5.2",
+    "firebase/php-jwt": "^6.4",
     "guzzlehttp/guzzle": "^7.3"
   },
   "require-dev": {
-    "alexeyshockov/guzzle-psalm-plugin": "^0.3.1",
+    "alexeyshockov/guzzle-psalm-plugin": "^1.0",
     "imbo/imbo-coding-standard": "^2.0",
-    "phpunit/phpunit": "^9.5",
-    "psalm/plugin-phpunit": "^0.16.1",
+    "phpunit/phpunit": "^10.0",
+    "psalm/plugin-phpunit": "^0.18.4",
     "slim/psr7": "^1.3",
     "slim/slim": "^4.7",
-    "symfony/process": "^5.2",
+    "symfony/process": "^6.2",
     "tuupola/slim-basic-auth": "^3.3",
-    "vimeo/psalm": "^4.7"
+    "vimeo/psalm": "^5.8"
   },
   "autoload": {
     "psr-4": {
@@ -49,6 +48,16 @@
     }
   },
   "scripts": {
+    "ci": [
+      "@phpunit",
+      "@behat",
+      "@sa",
+      "@cs"
+    ],
+    "test": [
+      "@phpunit",
+      "@behat"
+    ],
     "phpunit": "vendor/bin/phpunit",
     "phpunit:coverage": "vendor/bin/phpunit --coverage-html build/coverage",
     "behat": "vendor/bin/behat --strict",
