@@ -4,6 +4,7 @@ namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 use Firebase;
 use Imbo\BehatApiExtension\ArrayContainsComparator as Comparator;
 use InvalidArgumentException;
+use UnexpectedValueException;
 
 /**
  * Match a JWT token
@@ -12,24 +13,22 @@ class JWT
 {
     /**
      * Comparator for the array
-     *
-     * @var Comparator
      */
-    private $comparator;
+    private Comparator $comparator;
 
     /**
      * JWT tokens present in the response body
      *
-     * @var array<string, array{payload: array, secret: string}>
+     * @var array<string,array{payload:array,secret:string}>
      */
-    private $jwtTokens = [];
+    private array $jwtTokens = [];
 
     /**
      * Allowed algorithms for the JWT decoder
      *
-     * @var string[]
+     * @var array<string>
      */
-    protected $allowedAlgorithms = [
+    protected array $allowedAlgorithms = [
         'HS256',
         'HS384',
         'HS512',
@@ -65,12 +64,20 @@ class JWT
         }
 
         $token  = $this->jwtTokens[$name];
-        $result = (array) Firebase\JWT\JWT::decode($jwt, $token['secret'], $this->allowedAlgorithms);
 
-        if (!$this->comparator->compare($token['payload'], $result)) {
-            throw new InvalidArgumentException('JWT mismatch.');
+        foreach ($this->allowedAlgorithms as $algorithm) {
+            try {
+                $result = (array) Firebase\JWT\JWT::decode($jwt, new Firebase\JWT\Key($token['secret'], $algorithm));
+            } catch (UnexpectedValueException $e) {
+                // try next algorithm
+                continue;
+            }
+
+            if ($this->comparator->compare($token['payload'], $result)) {
+                return true;
+            }
         }
 
-        return true;
+        throw new InvalidArgumentException('JWT mismatch.');
     }
 }
