@@ -6,6 +6,7 @@ use Imbo\BehatApiExtension\Exception\ArrayContainsComparatorException;
 use Assert\AssertionFailedException as AssertionFailure;
 use Behat\Gherkin\Node\PyStringNode;
 use Behat\Gherkin\Node\TableNode;
+use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Request;
@@ -27,8 +28,6 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
 {
     /**
      * Guzzle client
-     *
-     * @var ClientInterface
      */
     protected ClientInterface $client;
 
@@ -41,10 +40,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Request instance
      *
      * The request instance will be created once the client is ready to send it.
-     *
-     * @var RequestInterface
      */
-    protected $request;
+    protected RequestInterface $request;
 
     /**
      * Request options
@@ -52,13 +49,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Options to send with the request.
      *
      * @var array{
-     *   auth: string[],
-     *   form_params: array<string, string|string[]>,
-     *   multipart: array<int, array{name: string, contents: string|resource}>,
-     *   query: array<string, mixed>
+     *   auth:array<string>,
+     *   form_params:array<string,string|array<string>>,
+     *   multipart:array<array{name:string,contents:string|resource,filename?:string}>,
+     *   query:array<string,mixed>
      * }
      */
-    protected $requestOptions = [
+    protected array $requestOptions = [
         'auth'        => [],
         'form_params' => [],
         'multipart'   => [],
@@ -69,60 +66,45 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Response instance
      *
      * The response object will be set once the request has been made.
-     *
-     * @var ?ResponseInterface
      */
-    protected $response;
+    protected ?ResponseInterface $response = null;
 
     /**
      * Instance of the comparator that handles matching of JSON
-     *
-     * @var ArrayContainsComparator
      */
-    protected $arrayContainsComparator;
+    protected ArrayContainsComparator $arrayContainsComparator;
 
     /**
      * Does HTTP method has been manually set
-     *
-     * @var bool
      */
-    protected $forceHttpMethod = false;
-
-    /**
-     * Request / response history for the Guzzle Client
-     *
-     * @var array{request: RequestInterface, response: ResponseInterface}[]
-     */
-    protected $clientHistory = [];
+    protected bool $forceHttpMethod = false;
 
     /**
      * Error message used when a required response instance if missing
-     *
-     * @var string
      */
-    protected $missingResponseError = 'The request has not been made yet, so no response object exists.';
+    protected string $missingResponseError = 'The request has not been made yet, so no response object exists.';
 
     /**
      * Set the client instance
      *
-     * @return self
+     * @throws InvalidArgumentException
      */
-    public function setClient(ClientInterface $client, string $baseUri)
+    public function initializeClient(array $config): static
     {
-        $this->client = $client;
-        $this->baseUri = $baseUri;
+        if (!array_key_exists('base_uri', $config) || !is_string($config['base_uri']) || '' === trim($config['base_uri'])) {
+            throw new InvalidArgumentException('base_uri is missing');
+        }
 
+        $this->baseUri = $config['base_uri'];
+        $this->client = new Client($config);
         $this->request = new Request('GET', $this->baseUri);
-
         return $this;
     }
 
     /**
      * Set the array contains comparator instance
-     *
-     * @return self
      */
-    public function setArrayContainsComparator(ArrayContainsComparator $comparator)
+    public function setArrayContainsComparator(ArrayContainsComparator $comparator): static
     {
         $this->arrayContainsComparator = $comparator;
 
@@ -136,11 +118,9 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $partName Multipart entry name
      * @throws InvalidArgumentException If the $path does not point to a file, an exception is
      *                                  thrown
-     * @return self
-     *
      * @Given I attach :path to the request as :partName
      */
-    public function addMultipartFileToRequest($path, $partName)
+    public function addMultipartFileToRequest(string $path, string $partName): static
     {
         if (!file_exists($path)) {
             throw new InvalidArgumentException(sprintf('File does not exist: "%s"', $path));
@@ -162,10 +142,9 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
     /**
      * Add an element to the multipart array
      *
-     * @param array{name: string, contents: resource|string, filename?: string} $part The part to add
-     * @return self
+     * @param array{name:string,contents:resource|string,filename?:string} $part The part to add
      */
-    private function addMultipartPart($part)
+    private function addMultipartPart(array $part): static
     {
         $this->requestOptions['multipart'][] = $part;
 
@@ -176,11 +155,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Add multipart form parameters to the request
      *
      * @param TableNode $table Table with name / value pairs
-     * @return self
      *
      * @Given the following multipart form parameters are set:
      */
-    public function setRequestMultipartFormParams(TableNode $table)
+    public function setRequestMultipartFormParams(TableNode $table): static
     {
         foreach ($this->getTableNodeHash($table) as $name => $value) {
             $this->addMultipartPart([
@@ -197,11 +175,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $username The username to authenticate with
      * @param string $password The password to authenticate with
-     * @return self
      *
      * @Given I am authenticating as :username with password :password
      */
-    public function setBasicAuth($username, $password)
+    public function setBasicAuth(string $username, string $password): static
     {
         $this->requestOptions['auth'] = [$username, $password];
 
@@ -217,12 +194,11 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $scope        The scope to authenticate in
      * @param string $clientId     The client_id to send
      * @param string $clientSecret Optional client_secret to send
-     * @return self
      *
      * @Given I get an OAuth token using password grant from :path with :username and :password in scope :scope using client ID :clientId
      * @Given I get an OAuth token using password grant from :path with :username and :password in scope :scope using client ID :clientId and client secret :clientSecret
      */
-    public function oauthWithPasswordGrantInScope($path, $username, $password, $scope, $clientId, $clientSecret = null)
+    public function oauthWithPasswordGrantInScope(string $path, string $username, string $password, string $scope, string $clientId, string $clientSecret = null): static
     {
         $this->requestOptions['form_params'] = array_filter([
             'grant_type'    => 'password',
@@ -279,11 +255,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $header The header name
      * @param string $value The header value
-     * @return self
      *
      * @Given the :header request header is :value
      */
-    public function setRequestHeader($header, $value)
+    public function setRequestHeader(string $header, string $value): static
     {
         $this->request = $this->request->withHeader($header, $value);
 
@@ -297,11 +272,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $header The header name
      * @param string $value The header value
-     * @return self
      *
      * @Given the :header request header contains :value
      */
-    public function addRequestHeader($header, $value)
+    public function addRequestHeader(string $header, string $value): static
     {
         $this->request = $this->request->withAddedHeader($header, $value);
 
@@ -312,13 +286,12 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Set form parameters
      *
      * @param TableNode $table Table with name / value pairs
-     * @return self
      *
      * @Given the following form parameters are set:
      */
-    public function setRequestFormParams(TableNode $table)
+    public function setRequestFormParams(TableNode $table): static
     {
-        /** @var array<string, array{name: string, value: string}> */
+        /** @var array<string,array{name:string,value:string}> */
         $rows = $table->getColumnsHash();
 
         foreach ($rows as $row) {
@@ -344,11 +317,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param resource|string|PyStringNode $string The content to set as the request body
      * @throws InvalidArgumentException If form_params or multipart is used in the request options
      *                                  an exception will be thrown as these can't be combined.
-     * @return self
      *
      * @Given the request body is:
      */
-    public function setRequestBody($string)
+    public function setRequestBody($string): static
     {
         if (!empty($this->requestOptions['multipart']) || !empty($this->requestOptions['form_params'])) {
             throw new InvalidArgumentException(
@@ -375,11 +347,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $path Path to a file
      * @throws InvalidArgumentException|RuntimeException
-     * @return self
      *
      * @Given the request body contains :path
      */
-    public function setRequestBodyToFileResource($path)
+    public function setRequestBodyToFileResource(string $path): static
     {
         if (!file_exists($path)) {
             throw new InvalidArgumentException(sprintf('File does not exist: "%s"', $path));
@@ -405,11 +376,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $secret The secret used to sign the token
      * @param PyStringNode $payload The payload for the JWT
      * @throws RuntimeException
-     * @return self
      *
      * @Given the response body contains a JWT identified by :name, signed with :secret:
      */
-    public function addJwtToken($name, $secret, PyStringNode $payload)
+    public function addJwtToken(string $name, string $secret, PyStringNode $payload): static
     {
         $jwtMatcher = $this->arrayContainsComparator->getMatcherFunction('jwt');
 
@@ -430,15 +400,14 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $name The name of the parameter
      * @param string|TableNode $value The value to add
-     * @return self
      *
      * @Given the query parameter :name is :value
      * @Given the query parameter :name is:
      */
-    public function setQueryStringParameter($name, $value)
+    public function setQueryStringParameter(string $name, string|TableNode $value): static
     {
         if ($value instanceof TableNode) {
-            /** @var string[] */
+            /** @var array<string> */
             $value = array_column($value->getHash(), 'value');
         }
 
@@ -451,11 +420,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Set multiple query parameters for the upcoming request
      *
      * @param TableNode $params The values to set
-     * @return self
      *
      * @Given the following query parameters are set:
      */
-    public function setQueryStringParameters(TableNode $params)
+    public function setQueryStringParameters(TableNode $params): static
     {
         foreach ($this->getTableNodeHash($params) as $name => $value) {
             $this->requestOptions['query'][$name] = $value;
@@ -469,12 +437,11 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $path The path to request
      * @param string $method The HTTP method to use
-     * @return self
      *
      * @When I request :path
      * @When I request :path using HTTP :method
      */
-    public function requestPath($path, $method = null)
+    public function requestPath(string $path, string $method = null): static
     {
         $this->setRequestPath($path);
 
@@ -495,7 +462,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response code is :code
      */
-    public function assertResponseCodeIs($code): bool
+    public function assertResponseCodeIs(int|string $code): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -522,7 +489,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response code is not :code
      */
-    public function assertResponseCodeIsNot($code): bool
+    public function assertResponseCodeIsNot(int|string $code): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -549,7 +516,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response reason phrase is :phrase
      */
-    public function assertResponseReasonPhraseIs($phrase): bool
+    public function assertResponseReasonPhraseIs(string $phrase): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -576,7 +543,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response reason phrase is not :phrase
      */
-    public function assertResponseReasonPhraseIsNot($phrase): bool
+    public function assertResponseReasonPhraseIsNot(string $phrase): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -602,7 +569,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response reason phrase matches :expression
      */
-    public function assertResponseReasonPhraseMatches($pattern): bool
+    public function assertResponseReasonPhraseMatches(string $pattern): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -633,7 +600,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response status line is :line
      */
-    public function assertResponseStatusLineIs($line): bool
+    public function assertResponseStatusLineIs(string $line): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -666,7 +633,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response status line is not :line
      */
-    public function assertResponseStatusLineIsNot($line): bool
+    public function assertResponseStatusLineIsNot(string $line): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -698,7 +665,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response status line matches :expression
      */
-    public function assertResponseStatusLineMatches($pattern): bool
+    public function assertResponseStatusLineMatches(string $pattern): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -743,7 +710,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response is :group
      */
-    public function assertResponseIs($group): bool
+    public function assertResponseIs(string $group): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -782,7 +749,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response is not :group
      */
-    public function assertResponseIsNot($group): bool
+    public function assertResponseIsNot(string $group): bool
     {
         try {
             $this->assertResponseIs($group);
@@ -810,7 +777,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the :header response header exists
      */
-    public function assertResponseHeaderExists($header): bool
+    public function assertResponseHeaderExists(string $header): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -836,7 +803,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the :header response header does not exist
      */
-    public function assertResponseHeaderDoesNotExist($header): bool
+    public function assertResponseHeaderDoesNotExist(string $header): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -863,7 +830,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the :header response header is :value
      */
-    public function assertResponseHeaderIs($header, $value): bool
+    public function assertResponseHeaderIs(string $header, string $value): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -896,7 +863,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the :header response header is not :value
      */
-    public function assertResponseHeaderIsNot($header, $value): bool
+    public function assertResponseHeaderIsNot(string $header, string $value): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -928,7 +895,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the :header response header matches :pattern
      */
-    public function assertResponseHeaderMatches($header, $pattern): bool
+    public function assertResponseHeaderMatches(string $header, string $pattern): bool
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -1033,7 +1000,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response body is a JSON array of length :length
      */
-    public function assertResponseBodyJsonArrayLength($length): bool
+    public function assertResponseBodyJsonArrayLength(int|string $length): bool
     {
         $this->requireResponse();
         $length = (int) $length;
@@ -1065,7 +1032,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response body is a JSON array with a length of at least :length
      */
-    public function assertResponseBodyJsonArrayMinLength($length): bool
+    public function assertResponseBodyJsonArrayMinLength(int|string $length): bool
     {
         $this->requireResponse();
 
@@ -1099,7 +1066,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @Then the response body is a JSON array with a length of at most :length
      */
-    public function assertResponseBodyJsonArrayMaxLength($length): bool
+    public function assertResponseBodyJsonArrayMaxLength(int|string $length): bool
     {
         $this->requireResponse();
 
@@ -1228,7 +1195,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         $contains = $this->jsonDecode((string) $contains);
 
         // Get the decoded response body and make sure it's decoded to an array
-        /** @var array<array-key, mixed> */
+        /** @var array<mixed> */
         $body = json_decode((string) json_encode($this->getResponseBody()), true);
 
         try {
@@ -1279,9 +1246,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Send the current request and set the response instance
      *
      * @throws RequestException
-     * @return self
      */
-    protected function sendRequest()
+    protected function sendRequest(): static
     {
         if (!empty($this->requestOptions['form_params']) && !$this->forceHttpMethod) {
             $this->setRequestMethod('POST');
@@ -1347,8 +1313,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $group The name of the group
      * @throws InvalidArgumentException
-     * @return array{min: int, max: int} An array with two keys, min and max, which represents the
-     *                                   min and max values for $group
+     * @return array{min:int,max:int} An array with two keys, min and max, which represents the
+     *                                min and max values for $group
      */
     protected function getResponseCodeGroupRange(string $group): array
     {
@@ -1387,21 +1353,16 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Get the "response group" based on a status code
      *
      * @param int $code The respose code
-     * @return string
      */
     protected function getResponseGroup(int $code): string
     {
-        if ($code >= 500) {
-            return 'server error';
-        } elseif ($code >= 400) {
-            return 'client error';
-        } elseif ($code >= 300) {
-            return 'redirection';
-        } elseif ($code >= 200) {
-            return 'success';
-        }
-
-        return 'informational';
+        return match (true) {
+            $code >= 500 => 'server error',
+            $code >= 400 => 'client error',
+            $code >= 300 => 'redirection',
+            $code >= 200 => 'success',
+            default => 'informational',
+        };
     }
 
     /**
@@ -1426,9 +1387,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Update the path of the request
      *
      * @param string $path The path to request
-     * @return self
      */
-    protected function setRequestPath(string $path)
+    protected function setRequestPath(string $path): static
     {
         $base = Utils::uriFor($this->baseUri);
         $uri = UriResolver::resolve($base, Utils::uriFor($path));
@@ -1444,9 +1404,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param bool $force Force the HTTP method. If set to false the method set CAN be
      *                       overridden (this occurs for instance when adding form parameters to the
      *                       request, and not specifying HTTP POST for the request)
-     * @return self
      */
-    protected function setRequestMethod(string $method, bool $force = true)
+    protected function setRequestMethod(string $method, bool $force = true): static
     {
         $this->request = $this->request->withMethod($method);
         $this->forceHttpMethod = $force;
@@ -1460,7 +1419,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @throws InvalidArgumentException
      * @return array<mixed>|stdClass
      */
-    protected function getResponseBody()
+    protected function getResponseBody(): array|stdClass
     {
         if (!$this->response) {
             throw new RuntimeException($this->missingResponseError);
@@ -1524,11 +1483,11 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * This method will effectively remove duplicates from TableNode
      *
      * @param TableNode $table
-     * @return array<string, string>
+     * @return array<string,string>
      */
     protected function getTableNodeHash(TableNode $table): array
     {
-        /** @var array<string, string> */
+        /** @var array<string,string> */
         return array_slice($table->getRowsHash(), 1);
     }
 }
