@@ -3,7 +3,6 @@ namespace Imbo\BehatApiExtension\Context;
 
 use Behat\Gherkin\Node\PyStringNode;
 use Behat\Gherkin\Node\TableNode;
-use GuzzleHttp\Client;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
@@ -52,26 +51,12 @@ function is_readable(string $path): bool
  */
 class ApiContextTest extends TestCase
 {
-    /** @var MockHandler */
-    private $mockHandler;
-
-    /** @var HandlerStack */
-    private $handlerStack;
-
-    /** @var array{request: Request, response: Response}[] */
-    private $historyContainer = [];
-
-    /** @var ApiContext */
-    private $context;
-
-    /** @var Client */
-    private $client;
-
-    /** @var MockObject&ArrayContainsComparator */
-    private $comparator;
-
-    /** @var string */
-    private $baseUri = 'http://localhost:9876';
+    private ApiContext $context;
+    private MockHandler $mockHandler;
+    private HandlerStack $handlerStack;
+    private MockObject&ArrayContainsComparator $comparator;
+    /** @var array<array{request:Request,response:Response}> */
+    private array $historyContainer = [];
 
     public function setUp(): void
     {
@@ -80,21 +65,20 @@ class ApiContextTest extends TestCase
         $this->mockHandler = new MockHandler();
         $this->handlerStack = HandlerStack::create($this->mockHandler);
         $this->handlerStack->push(Middleware::history($this->historyContainer));
-        $this->client = new Client([
-            'handler' => $this->handlerStack,
-            'base_uri' => $this->baseUri,
-        ]);
         $this->comparator = $this->createMock(ArrayContainsComparator::class);
 
         $this->context = new ApiContext();
-        $this->context->setClient($this->client, 'http://base');
-        $this->context->setArrayContainsComparator($this->comparator);
+        $this->context->initializeClient([
+            'handler' => $this->handlerStack,
+            'base_uri' => 'http://localhost:9876',
+        ]);
+        $this->assertSame($this->context, $this->context->setArrayContainsComparator($this->comparator));
     }
 
     /**
-     * @return array{method: string}[]
+     * @return array<array{method:string}>
      */
-    public function getHttpMethods(): array
+    public static function getHttpMethods(): array
     {
         return [
             ['method' => 'GET'],
@@ -108,29 +92,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{filePath: string, method: string, expectedMimeType: string}[]
+     * @return array<array{code:int,others:array<int>}>
      */
-    public function getFilesAndMimeTypes(): array
-    {
-        return [
-            [
-                'filePath'         => __FILE__,
-                'method'           => 'POST',
-                'expectedMimeType' => 'text/x-php',
-            ],
-            [
-                'filePath'         => __FILE__,
-                'method'           => 'POST',
-                'expectedMimeType' => 'foo/bar',
-                'mimeType'         => 'foo/bar',
-            ],
-        ];
-    }
-
-    /**
-     * @return array{code: int, others: int[]}[]
-     */
-    public function getResponseCodes(): array
+    public static function getResponseCodes(): array
     {
         return [
             [
@@ -153,9 +117,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{group: string, codes: int[]}[]
+     * @return array<array{group:string,codes:array<int>}>
      */
-    public function getGroupAndResponseCodes(): array
+    public static function getGroupAndResponseCodes(): array
     {
         return [
             [
@@ -182,9 +146,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return int[][]
+     * @return array<array<int>>
      */
-    public function getInvalidHttpResponseCodes(): array
+    public static function getInvalidHttpResponseCodes(): array
     {
         return [
             [99],
@@ -193,9 +157,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{body: int[], lengthToUse: int, willFail: bool}[]
+     * @return array<array{body:array<int>,lengthToUse:int,willFail:bool}>
      */
-    public function getResponseBodyArrays(): array
+    public static function getResponseBodyArrays(): array
     {
         return [
             [
@@ -222,9 +186,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{body: int[], min: int}[]
+     * @return array<array{body:array<int>,min:int}>
      */
-    public function getResponseBodyArraysForAtLeast(): array
+    public static function getResponseBodyArraysForAtLeast(): array
     {
         return [
             [
@@ -247,9 +211,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{body: int[], max: int}[]
+     * @return array<array{body:array<int>,max:int}>
      */
-    public function getResponseBodyArraysForAtMost(): array
+    public static function getResponseBodyArraysForAtMost(): array
     {
         return [
             [
@@ -272,9 +236,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{code: int, phrase: string}[]
+     * @return array<array{code:int,phrase:string}>
      */
-    public function getResponseCodesAndReasonPhrases(): array
+    public static function getResponseCodesAndReasonPhrases(): array
     {
         return [
             200 => [
@@ -297,9 +261,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{baseUri: string, path: string, fullUri: string}[]
+     * @return array<array{baseUri:string,path:string,fullUri:string}>
      */
-    public function getUris(): array
+    public static function getUris(): array
     {
         return [
             // The first six sets are from http://docs.guzzlephp.org/en/latest/quickstart.html (2016-12-12)
@@ -360,9 +324,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{responseCode: int, actualGroup: string, expectedGroup: string}[]
+     * @return array<array{responseCode:int,actualGroup:string,expectedGroup:string}>
      */
-    public function getDataForResponseGroupFailures(): array
+    public static function getDataForResponseGroupFailures(): array
     {
         return [
             [
@@ -394,9 +358,9 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array<array-key, array{data: string|PyStringNode, expected: string}>
+     * @return array<array{data:string|PyStringNode,expected:string}>
      */
-    public function getRequestBodyValues(): array
+    public static function getRequestBodyValues(): array
     {
         return [
             [
@@ -407,6 +371,19 @@ class ApiContextTest extends TestCase
                 'data' => new PyStringNode(['some string'], 1),
                 'expected' => 'some string',
             ],
+        ];
+    }
+
+    /**
+     * @return array<array{httpMethod:string}>
+     */
+    public static function getHttpMethodsForFormParametersTest(): array
+    {
+        return [
+            ['httpMethod' => 'PUT'],
+            ['httpMethod' => 'POST'],
+            ['httpMethod' => 'PATCH'],
+            ['httpMethod' => 'DELETE'],
         ];
     }
 
@@ -631,19 +608,6 @@ class ApiContextTest extends TestCase
     }
 
     /**
-     * @return array{httpMethod: string}[]
-     */
-    public function getHttpMethodsForFormParametersTest(): array
-    {
-        return [
-            ['httpMethod' => 'PUT'],
-            ['httpMethod' => 'POST'],
-            ['httpMethod' => 'PATCH'],
-            ['httpMethod' => 'DELETE'],
-        ];
-    }
-
-    /**
      * @dataProvider getHttpMethodsForFormParametersTest
      * @covers ::setRequestFormParams
      * @covers ::sendRequest
@@ -754,9 +718,8 @@ BAR;
     /**
      * @dataProvider getRequestBodyValues
      * @covers ::setRequestBody
-     * @param string|PyStringNode $data
      */
-    public function testCanSetRequestBodyToAString($data, string $expected): void
+    public function testCanSetRequestBodyToAString(string|PyStringNode $data, string $expected): void
     {
         $this->mockHandler->append(new Response());
         $this->context->setRequestBody($data);
@@ -769,16 +732,16 @@ BAR;
 
     /**
      * @dataProvider getUris
-     * @covers ::setClient
+     * @covers ::initializeClient
      * @covers ::setRequestPath
      */
     public function testResolvesFilePathsCorrectlyWhenAttachingFilesToTheRequestBody(string $baseUri, string $path, string $fullUri): void
     {
         // Set a new client with the given base_uri (and not the one used in setUp())
-        $this->assertSame($this->context, $this->context->setClient(new Client([
+        $this->assertSame($this->context, $this->context->initializeClient([
             'handler' => $this->handlerStack,
             'base_uri' => $baseUri,
-        ]), $baseUri));
+        ]));
 
         $this->mockHandler->append(new Response());
         $this->assertSame($this->context, $this->context->setRequestBodyToFileResource(__FILE__));
@@ -799,6 +762,7 @@ BAR;
         $payload = ['some' => 'data'];
         $secret = 'secret';
 
+        /** @var MockObject&JWT */
         $matcher = $this->createMock(JWT::class);
         $matcher
             ->expects($this->once())
@@ -943,7 +907,7 @@ BAR;
      * @covers ::assertResponseIs
      * @covers ::requireResponse
      * @covers ::getResponseCodeGroupRange
-     * @param int[] $codes
+     * @param array<int> $codes
      */
     public function testCanAssertWhichGroupTheResponseIsIn(string $group, array $codes): void
     {
@@ -958,7 +922,7 @@ BAR;
      * @dataProvider getGroupAndResponseCodes
      * @covers ::assertResponseIsNot
      * @covers ::assertResponseIs
-     * @param int[] $codes
+     * @param array<int> $codes
      */
     public function testCanAssertWhichGroupTheResponseIsNotIn(string $group, array $codes): void
     {
@@ -974,9 +938,7 @@ BAR;
             $this->mockHandler->append(new Response($code));
             $this->context->requestPath('/some/path');
 
-            foreach (array_filter($groups, function (string $g) use ($group): bool {
-                return $g !== $group;
-            }) as $g) {
+            foreach (array_filter($groups, fn (string $g): bool => $g !== $group) as $g) {
                 // Assert that the response is not in any of the other groups
                 $this->assertTrue($this->context->assertResponseIsNot($g));
             }
@@ -1183,7 +1145,7 @@ BAR;
      * @dataProvider getResponseBodyArrays
      * @covers ::assertResponseBodyJsonArrayLength
      * @covers ::getResponseBodyArray
-     * @param int[] $body
+     * @param array<int> $body
      */
     public function testCanAssertThatTheArrayInTheResponseBodyHasACertainLength(array $body, int $lengthToUse, bool $willFail): void
     {
@@ -1207,7 +1169,7 @@ BAR;
      * @dataProvider getResponseBodyArraysForAtLeast
      * @covers ::assertResponseBodyJsonArrayMinLength
      * @covers ::getResponseBody
-     * @param int[] $body
+     * @param array<int> $body
      */
     public function testCanAssertTheMinLengthOfAnArrayInTheResponseBody(array $body, int $min): void
     {
@@ -1220,7 +1182,7 @@ BAR;
      * @dataProvider getResponseBodyArraysForAtMost
      * @covers ::assertResponseBodyJsonArrayMaxLength
      * @covers ::getResponseBody
-     * @param int[] $body
+     * @param array<int> $body
      */
     public function testCanAssertTheMaxLengthOfAnArrayInTheResponseBody(array $body, int $max): void
     {
@@ -1245,7 +1207,7 @@ BAR;
             ->with(['bar' => 'foo', 'foo' => 'bar'], ['foo' => 'bar', 'bar' => 'foo'])
             ->willReturn(true);
 
-        $this->context->assertResponseBodyContainsJson(new PyStringNode(['{"bar":"foo","foo":"bar"}'], 1));
+        $this->assertTrue($this->context->assertResponseBodyContainsJson(new PyStringNode(['{"bar":"foo","foo":"bar"}'], 1)));
     }
 
     /**
