@@ -27,15 +27,15 @@ $app->any('/', function (Request $request, Response $response): Response {
         'string' => 'value',
         'integer' => 42,
         'float' => 4.2,
-        'boolean true' => true,
-        'boolean false' => false,
+        'bool true' => true,
+        'bool false' => false,
         'list' => [1, 2, 3, [1], ['foo' => 'bar']],
         'sub' => [
             'string' => 'value',
             'integer' => 42,
             'float' => 4.2,
-            'boolean true' => true,
-            'boolean false' => false,
+            'bool true' => true,
+            'bool false' => false,
             'list' => [1, 2, 3, [1], ['foo' => 'bar']],
         ],
         'types' => [
@@ -43,7 +43,7 @@ $app->any('/', function (Request $request, Response $response): Response {
             'integer' => 123,
             'double' => 1.23,
             'array' => [1, '2', 3],
-            'boolean' => true,
+            'bool' => true,
             'null' => null,
             'scalar' => '123',
         ],
