@@ -3,7 +3,6 @@ namespace Imbo\BehatApiExtension\Context\Initializer;
 
 use Behat\Behat\Context\Context;
 use Behat\Behat\Context\Initializer\ContextInitializer;
-use GuzzleHttp\Client;
 use Imbo\BehatApiExtension\Context\ApiClientAwareContext;
 
 /**
@@ -14,19 +13,19 @@ use Imbo\BehatApiExtension\Context\ApiClientAwareContext;
 class ApiClientAwareInitializer implements ContextInitializer
 {
     /**
-     * @var array{base_uri?: string} Guzzle client configuration array
+     * @var array Guzzle client configuration array
      * @see http://docs.guzzlephp.org/ Check out the Guzzle docs for a complete overview of available configuration parameters
      */
-    private array $guzzleConfig = [];
+    private array $config = [];
 
     /**
      * Class constructor
      *
-     * @param array{base_uri?: string} $guzzleConfig Guzzle client configuration array
+     * @param array $config Guzzle client configuration array
      */
-    public function __construct(array $guzzleConfig)
+    public function __construct(array $config)
     {
-        $this->guzzleConfig = $guzzleConfig;
+        $this->config = $config;
     }
 
     /**
@@ -37,10 +36,7 @@ class ApiClientAwareInitializer implements ContextInitializer
     public function initializeContext(Context $context): void
     {
         if ($context instanceof ApiClientAwareContext) {
-            $context->setClient(
-                new Client($this->guzzleConfig),
-                $this->guzzleConfig['base_uri'] ?? '',
-            );
+            $context->initializeClient($this->config);
         }
     }
 }
