@@ -8,18 +8,8 @@ use Exception;
  */
 class ArrayContainsComparatorException extends AssertionFailedException
 {
-    /**
-     * Class constructor
-     *
-     * @param string $message
-     * @param int $code
-     * @param Exception $previous
-     * @param mixed $needle
-     * @param mixed $haystack
-     */
-    public function __construct(string $message, int $code = 0, Exception $previous = null, $needle = null, $haystack = null)
+    public function __construct(string $message, int $code = 0, Exception $previous = null, mixed $needle = null, mixed $haystack = null)
     {
-        // Format the error message
         $message .= PHP_EOL . PHP_EOL . sprintf(
             <<<MESSAGE
 ================================================================================
