@@ -9,9 +9,9 @@ use PHPUnit\Framework\TestCase;
 class ArrayContainsComparatorExceptionTest extends TestCase
 {
     /**
-     * @return array<array-key, array{message: string, needle: array<string, string>, haystack: array<string, string>, formattedMessage: string}>
+     * @return array<array{message:string,needle:array<string,string>,haystack:array<string,string>,formattedMessage:string}>
      */
-    public function getExceptionData(): array
+    public static function getExceptionData(): array
     {
         return [
             'with no needle / haystack' => [
@@ -60,10 +60,6 @@ MESSAGE
     /**
      * @dataProvider getExceptionData
      * @covers ::__construct
-     * @param string $message
-     * @param array<string, string> $needle
-     * @param array<string, string> $haystack
-     * @param string $formattedMessage
      */
     public function testCanProperlyFormatErrorMessages(string $message, array $needle, array $haystack, string $formattedMessage): void
     {
